--- src/components/VideoPlayer.vue
@@ -1,47 +1,0 @@
-<template>
-  <div>
-    <video ref="videoPlayer" class="video-js"></video>
-  </div>
-</template>
-
-<script>
-import videojs from 'video.js'
-
-export default {
-  name: "VideoPlayer",
-  props: {
-    options: {
-      type: Object,
-      default() {
-        return {};
-      }
-    }
-  },
-  data() {
-    return {
-      player: null
-    }
-  },
-  mounted() {
-    this.player = videojs(
-        this.$refs.videoPlayer,
-        this.options,
-        function onPlayerReady() {
-          console.log('onPlayerReady', this);
-        },
-    )
-
-    // Auto plays on load
-    // this.player.play();
-  },
-  beforeDestroy() {
-    if (this.player) {
-      this.player.dispose();
-    }
-  }
-}
-</script>
-
-<style scoped>
-@import "~video.js/dist/video-js.min.css";
-</style>
